Revert "[0.7.0] Run LSP in wasm context for compatibility with vscode for web…"
This reverts commit 7b9317754b7e554ffe8d07abc08e234e6613ab6c.
