+++ .gitignore
@@ -2,4 +2,3 @@ node_modules
 *.vsix
 .vscode/settings.json
 package_lock.json
-dist
